+++ src/app/app.module.ts
@@ -8,11 +8,7 @@ import { ContentModule } from '@alfresco/adf-content-services';
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { appRoutes } from './app.routes';
 import { FileViewComponent } from './file-view/file-view.component';
-
-// App components
 import { AppComponent } from './app.component';
-import { HomeComponent } from './home/home.component';
-import { LoginComponent } from './login/login.component';
 import { DocumentsComponent } from './documents/documents.component';
 import { AppLayoutComponent } from './app-layout/app-layout.component';
 
@@ -34,7 +30,7 @@ registerLocaleData(localeEs);
     AppLayoutComponent,
     DocumentsComponent
   ],
-  declarations: [AppComponent, HomeComponent, LoginComponent, FileViewComponent],
+  declarations: [AppComponent, FileViewComponent],
   providers: [provideTranslations('app', 'resources')],
   bootstrap: [AppComponent]
 })