+++ src/app/home/home.component.ts
@@ -1,6 +1,9 @@
 import { Component } from '@angular/core';
+import { TranslateModule } from '@ngx-translate/core';
 
 @Component({
+  standalone: true,
+  imports: [TranslateModule],
   selector: 'app-home',
   templateUrl: './home.component.html'
 })