+++ src/app/login/login.component.ts
@@ -1,7 +1,12 @@
-import { Component } from '@angular/core';
+import { AppConfigModule, LoginModule } from '@alfresco/adf-core';
+import { Component, ViewEncapsulation } from '@angular/core';
+import { TranslateModule } from '@ngx-translate/core';
 
 @Component({
+  standalone: true,
+  imports: [LoginModule, AppConfigModule, TranslateModule],
   selector: 'app-login',
-  templateUrl: './login.component.html'
+  templateUrl: './login.component.html',
+  encapsulation: ViewEncapsulation.None
 })
 export class LoginComponent {}