+++ CHANGELOG.md
@@ -15,8 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New simplified `Sidenav` component
 - `Outlined` material icons font by default (see [material-icons](https://github.com/marella/material-icons))
-- Migrate `AppLayout` component to a Standalone Angular component
-- Migrate `Documents` component to a Standalone Angular component
+- Migrate all components to Standalone ones
 - `Documents`: enable column resizing feature by default
 
 ## [2.0.0] - 2024-01-26